+++ sim-inventory/src/main/java/org/ostelco/SimAdministrationApplication.kt
@@ -1,6 +1,5 @@
 package org.ostelco
 
-import com.fasterxml.jackson.annotation.JsonProperty
 import io.dropwizard.Application
 import io.dropwizard.jdbi.DBIFactory
 import io.dropwizard.setup.Bootstrap
@@ -73,7 +72,7 @@ class SimAdministrationApplication : Application<SimAdministrationAppConfigurati
         environment.jersey().register(OpenApiResource()
                 .openApiConfiguration(oasConfig))
 
-        environment.jersey().register(EsimInventoryResource(simInventoryDAO))
+        environment.jersey().register(SimInventoryResource(simInventoryDAO))
     }
 
     companion object {
@@ -104,217 +103,4 @@ class AppExceptionMapper : ExceptionMapper<Es2Exception> {
 }
 */
 
-///
-///  The web resource using the protocol domain model.
-///
 
-@Path("/ostelco/sim-inventory/{hlr}/")
-class EsimInventoryResource(val dao: SimInventoryDAO) {
-
-    private fun <T> assertNonNull(v: T?): T {
-        if (v == null) {
-            throw WebApplicationException(Response.Status.NOT_FOUND)
-        } else {
-            return v
-        }
-    }
-
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Path("iccid/{iccid}")
-    @GET
-    fun findByIccid(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
-
-        val sim = assertNonNull(dao.getSimProfileByIccid(iccid))
-        assertHlrsEqual(hlr, sim.hlrId)
-        return sim
-    }
-
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Path("imsi/{imsi}")
-    @GET
-    fun findByImsi(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("imsi") imsi: String): SimEntry {
-        val sim = assertNonNull(dao.getSimProfileByImsi(imsi))
-        assertHlrsEqual(hlr, sim.hlrId)
-        return sim
-    }
-
-
-    @Path("msisdn/{msisdn}")
-    @GET
-    fun findByMsisdn(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("msisdn") msisdn: String): SimEntry {
-        return assertNonNull(dao.getSimProfileByMsisdn(msisdn))
-    }
-
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-
-    @Path("msisdn/{msisdn}/allocate-next-free")
-    @GET
-    fun allocateNextFree(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("msisdn") msisdn: String): SimEntry {
-        return assertNonNull(dao.allocateNextFreeSimForMsisdn(hlr, msisdn))
-    }
-
-
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Path("/iccid/{iccid}/activate/all")
-    @GET
-    fun activateByIccid(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
-        val sim = activateHlrProfileByIccid(hlr, iccid)
-        assertHlrsEqual(hlr, sim.hlrId)
-        if (sim.smdpplus != null) {
-            return activateEsimProfileByIccid(hlr, iccid)
-        } else {
-            return sim
-        }
-    }
-
-    fun assertHlrsEqual(hlr1: String, hlr2: String) {
-        if (!hlr1.equals(hlr2)) {
-            // XXX log.error("Attempt to impersonate HLR.  '$hlr1', '$hlr2'")
-            throw WebApplicationException(Response.Status.BAD_REQUEST)
-        }
-    }
-
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Path("/iccid/{iccid}/activate/hlr")
-    @GET
-    fun activateHlrProfileByIccid(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
-
-        val simEntry = assertNonNull(dao.getSimProfileByIccid(iccid))
-        assertHlrsEqual(hlr, simEntry.hlrId)
-        val hlrAdapter = assertNonNull(dao.getHlrAdapterByName(hlr))
-
-        try {
-            hlrAdapter.activate(simEntry)
-        } catch (e: Exception) {
-            throw WebApplicationException(Response.Status.BAD_REQUEST)
-        }
-        return assertNonNull(dao.setActivatedInHlr(simEntry.id!!))
-    }
-
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Path("/iccid/{iccid}/activate/esim")
-    @GET
-    fun activateEsimProfileByIccid(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
-        val simEntry = assertNonNull(dao.getSimProfileByIccid(iccid))
-        assertHlrsEqual(hlr, simEntry.hlrId)
-
-        if (simEntry.smdpplus == null) {
-            throw WebApplicationException(Response.Status.BAD_REQUEST)
-        }
-        val smdpPlusAdpter = assertNonNull(dao.getSmdpPlusAdapterByName(simEntry.smdpplus))
-
-        try {
-            smdpPlusAdpter.activateEntry(simEntry)
-        } catch (e: Exception) {
-            throw WebApplicationException(Response.Status.BAD_REQUEST)
-        }
-        return assertNonNull(dao.setActivatedInSmdpPlus(simEntry.id!!))
-    }
-
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Path("/iccid/{iccid}/deactivate/hlr")
-    @GET
-    fun deactrivateByIccid(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
-        val simEntry = assertNonNull(dao.getSimProfileByIccid(iccid))
-
-        assertHlrsEqual(hlr, simEntry.hlrId)
-        val hlrAdapter = assertNonNull(dao.getHlrAdapterByName(hlr))
-
-        hlrAdapter.deactivate(simEntry)
-        dao.setActivatedInHlr(simEntry.id!!)
-        return dao.getSimProfileById(simEntry.id!!)
-    }
-
-    @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.TEXT_PLAIN)
-    @Throws(IOException::class)
-    @Path("/import-batch/sim-profile-vendor/{profilevendor}")
-    @PUT
-    fun importBatch(
-            @NotEmpty @PathParam("hlr") hlr: String,
-            @NotEmpty @PathParam("profilevendor") profilevendor: String,
-            csvInputStream: InputStream): SimImportBatch {
-
-
-        val  pvp  =
-                assertNonNull(dao.getProfilevendorByName(profilevendor))
-
-        val hlrAdapter = assertNonNull(dao.getHlrAdapterByName(hlr))
-
-        if (!pvp.isAuthorizedForHlr(hlr)) {
-            throw WebApplicationException(Response.Status.BAD_REQUEST)
-        }
-
-
-        return dao.importSims(
-                importer = "importer",
-                hlr = hlr,
-                profileVendor = profilevendor,
-                csvInputStream = csvInputStream)
-    }
-}
-
-
-/**
- * An adapter that can connect to HLR entries and activate/deactivate
- * individual SIM profiles.
- */
-data class HlrAdapter(
-        @JsonProperty("id") val id: Long,
-        @JsonProperty("name") val name: String) {
-
-    /**
-     * Will connect to the HLR and then activate the profile, so that when
-     * a VLR asks the HLR for the an authentication triplet, then the HLR
-     * will know that it should give an answer.
-     */
-    fun activate(simEntry: SimEntry) {
-        // XXX TBD
-    }
-
-    fun deactivate(simEntry: SimEntry) {
-        // XXX TBD
-    }
-}
-
-
-/**
- * An adapter that can connect to HLR entries and activate/deactivate
- * individual SIM profiles.
- */
-data class SmdpPlusAdapter(
-        @JsonProperty("id") val id: Long,
-        @JsonProperty("name") val name: String) {
-
-    /**
-     * Will connect to the SM-DP+  and then activate the profile, so that when
-     * user equpiment tries to download a profile, it will get a profile to
-     * download.
-     */
-    fun activateEntry(simEntry: SimEntry) {
-        // XXX TBD
-    }
-}
